Split i3 config, fix polybar multimonitor
i3 config is now split between two files - one with most of the custom config,
one with the default stuff. Polybar launching has been reworked to get the tray
to consistently start on the appropriate screen, although I'm sure I'll have to
rework the script to work with the laptop.


Former-commit-id: e3022974296ceb3fd732b1249f53efb5b021e01a